+++ src/wrappers/mj_editing.rs
@@ -5,6 +5,16 @@ use std::marker::PhantomData;
 use std::path::Path;
 use std::ptr;
 
+#[macro_use]
+mod utility;
+use utility::*;
+
+mod traits;
+pub use traits::*;
+
+mod default;
+pub use default::*;
+
 use super::mj_model::{
     MjModel, MjtObj, MjtGeom, MjtJoint, MjtCamLight,
     MjtLightType, MjtSensor, MjtDataType, MjtGain,
@@ -33,311 +43,8 @@ pub type MjtOrientation = mjtOrientation;
 /// Type of built-in procedural texture.
 pub type MjtBuiltin = mjtBuiltin;
 
+/// Mark type for procedural textures.
 pub type MjtMark = mjtMark;
-
-
-/// Represents all the types that [`MjSpec`] supports.
-/// This is pre-implemented for all the specification types and is not
-/// meant to be implemented by the user.
-pub trait SpecItem: Sized {
-    /// Returns the internal element struct.
-    /// The element struct is the C++ implementation of the
-    /// actual item, which is hidden to the user, but is needed
-    /// in some functions.
-    /// 
-    /// # SAFETY
-    /// This borrows immutably, but returns a mutable pointer. This is done to overcome MJS's wrong
-    /// use of mutable pointers in functions, such as [`mjs_getName`].
-    unsafe fn element_pointer(&self) -> *mut mjsElement;
-
-    /// Same as [`SpecItem::element_pointer`], but with a mutable borrow.
-    unsafe fn element_mut_pointer(&mut self) -> *mut mjsElement {
-        unsafe { self.element_pointer() }
-    }
-
-    /// Returns the item's name.
-    fn name(&self) -> &str {
-        read_mjs_string( unsafe { mjs_getName(self.element_pointer()).as_ref().unwrap() } )
-    }
-
-    /// Set a new name.
-    fn set_name(&mut self, name: &str) {
-        let cstr = CString::new(name).unwrap();  // always has valid UTF-8
-        unsafe { mjs_setName(self.element_mut_pointer(), cstr.as_ptr()) };
-    }
-
-    /// Returns the used default.
-    fn default(&self) -> MjsDefault {
-        MjsDefault(unsafe { mjs_getDefault(self.element_pointer()) }, PhantomData)
-    }
-
-    /// Make the item inherit properties from `default`.
-    /// # Errors
-    /// Returns a [`ErrorKind::NotFound`] when the default with the `class_name` doesn't exist.
-    fn set_default(&mut self, class_name: &str) -> Result<(), Error> {
-        /* Workaround to pass the borrow checker (we use the existing borrow) */
-        let cname = CString::new(class_name).unwrap();  // class_name is always valid UTF-8.
-        let element = unsafe { self.element_mut_pointer() };
-        let spec = unsafe { mjs_getSpec(element) };
-        let default = unsafe { mjs_findDefault(spec, cname.as_ptr()) };
-        if default.is_null() {
-            return Err(Error::new(ErrorKind::NotFound, "class doesn't exist"));
-        }
-
-        unsafe { mjs_setDefault(self.element_mut_pointer(), default); }
-        Ok(())
-    }
-
-    /// Delete the item.
-    fn delete(mut self) -> Result<(), Error> {
-        let element = unsafe { self.element_mut_pointer() };
-        let spec = unsafe { mjs_getSpec(element) };
-        let result = unsafe { mjs_delete(spec, element) };
-        match result {
-            0 => Ok(()),
-            _ => Err(Error::new(ErrorKind::Other, unsafe { CStr::from_ptr(mjs_getError(spec)).to_string_lossy().into_owned() }))
-        }
-    }
-}
-
-/***************************
-** Utility functions
-***************************/
-/// Reads a MJS string (C++) as a `&str`.
-fn read_mjs_string(string: &mjString) -> &str {
-    unsafe {
-        let ptr = mjs_getString(string);
-        CStr::from_ptr(ptr).to_str().unwrap()  // can't be invalid UTF-8.
-    }
-}
-
-/// Writes to a `destination` MJS string (C++) from a `source` `&str`.
-fn write_mjs_string(source: &str, destination: &mut mjString) {
-    unsafe {
-        let c_source = CString::new(source).unwrap();  // can't be invalid UTF-8.
-        mjs_setString(destination, c_source.as_ptr());
-    }
-}
-
-/// Reads as MJS double vector (C++) as a `&\[f64\]`.
-fn read_mjs_vec_f64(array: &mjDoubleVec) -> &[f64] {
-    let mut userdata_length = 0;
-    unsafe {
-        let ptr_arr = mjs_getDouble(array, &mut userdata_length);
-        if ptr_arr.is_null() {
-            return &[];
-        }
-
-        std::slice::from_raw_parts(ptr_arr, userdata_length as usize)
-    }
-}
-
-/// Writes as MJS double vector (C++) from a `source` to `destination`.
-fn write_mjs_vec_f64(source: &[f64], destination: &mut mjDoubleVec) {
-    unsafe {
-        mjs_setDouble(destination, source.as_ptr(), source.len() as i32);
-    }
-}
-
-
-/// Writes as MJS float vector (C++) from a `source` to `destination`.
-fn write_mjs_vec_f32(source: &[f32], destination: &mut mjFloatVec) {
-    unsafe {
-        mjs_setFloat(destination, source.as_ptr(), source.len() as i32);
-    }
-}
-
-/// Writes as MJS int vector (C++) from a `source` to `destination`.
-fn write_mjs_vec_int(source: &[i32], destination: &mut mjIntVec) {
-    unsafe {
-        mjs_setInt(destination, source.as_ptr(), source.len() as i32);
-    }
-}
-
-/// Split `source` to entries and copy to `destination` (C++).
-fn write_mjs_vec_string(source: &str, destination: &mut mjStringVec) {
-    let c_source = CString::new(source).unwrap();  // can't be invalid UTF-8.
-    unsafe {
-        mjs_setStringVec(destination, c_source.as_ptr());
-    }
-}
-
-/// Split `source` to entries and append to `destination` (C++).
-fn append_mjs_vec_string(source: &str, destination: &mut mjStringVec) {
-    let c_source = CString::new(source).unwrap();  // can't be invalid UTF-8.
-    unsafe {
-        mjs_appendString(destination, c_source.as_ptr());
-    }
-}
-
-/***************************
-** Helper macros
-***************************/
-/// Creates an `add_ $name` method for adding new elements of $name into the body / spec.
-/// Also sets the default as null();
-macro_rules! add_x_method {
-    ($($name:ident),*) => {paste::paste! {
-        $(
-            /* Without default */
-            #[doc = concat!("Add and return a child ", stringify!($name), ".")]
-            pub fn [<add_ $name>](&mut self) -> [<Mjs $name:camel>] {
-                let ptr = unsafe { [<mjs_add $name:camel>](self.0, ptr::null()) };
-                [<Mjs $name:camel>](ptr, PhantomData)
-            }
-        )*
-    }};
-}
-
-/// Creates an `add_ $name` method for adding new elements of $name into the body / spec.
-/// This is for methods that don't accept a default.
-macro_rules! add_x_method_no_default {
-    ($($name:ident),*) => {paste::paste! {
-        $(
-            #[doc = concat!("Add and return a child ", stringify!($name), ".")]
-            pub fn [<add_ $name>](&mut self) -> [<Mjs $name:camel>] {
-                let ptr = unsafe { [<mjs_add $name:camel>](self.0) };
-                [<Mjs $name:camel>](ptr, PhantomData)
-            }
-        )*
-    }};
-}
-
-
-/// Creates an `get_ $name` method for finding items in spec / body.
-/// Also sets the default as null();
-macro_rules! find_x_method {
-    ($($item:ident),*) => {paste::paste! {
-        $(
-            #[doc = concat!("Obtain a reference to the ", stringify!($item), " with the given `name`.")]
-            pub fn $item(&self, name: &str) -> Option<[<Mjs $item:camel>]> {
-                let c_name = CString::new(name).unwrap();
-                unsafe {
-                    let ptr = mjs_findElement(self.0, MjtObj::[<mjOBJ_ $item:upper>], c_name.as_ptr());
-                    if ptr.is_null() {
-                        None
-                    }
-                    else {
-                        Some([<Mjs $item:camel>]([<mjs_as $item:camel>](ptr), PhantomData))
-                    }
-                }
-            }
-        )*
-    }};
-}
-
-/// Same as [`find_x_method`], but for types that have corresponding methods (instead of `mjs_findElement`).
-macro_rules! find_x_method_direct {
-    ($($item:ident),*) => {paste::paste!{
-        $(
-            #[doc = concat!("Obtain a reference to the ", stringify!($item), " with the given `name`.")]
-            pub fn $item(&self, name: &str) -> Option<[<Mjs $item:camel>]> {
-                let c_name = CString::new(name).unwrap();
-                unsafe {
-                    let ptr = [<mjs_find $item:camel>](self.0, c_name.as_ptr());
-                    if ptr.is_null() {
-                        None
-                    }
-                    else {
-                        Some([<Mjs $item:camel>](ptr, PhantomData))
-                    }
-                }
-            }
-        )*
-    }};
-}
-
-
-/// Creates a wrapper around a mjs$ffi_name item. It also implements the methods: `ffi()`, `ffi_mut()`
-/// and traits: [`SpecItem`], [`Sync`], [`Send`].
-macro_rules! mjs_wrapper {
-    ($ffi_name:ident) => {paste::paste!{
-        #[doc = concat!(stringify!($ffi_name), " specification. This acts as a safe reference to a FFI type [`", stringify!([<mjs $ffi_name>]), "`] internally.")]
-        pub struct [<Mjs $ffi_name>]<'s>(*mut [<mjs $ffi_name>], PhantomData<&'s mut ()>);  // the lifetime belongs to the parent
-
-        impl [<Mjs $ffi_name>]<'_> {
-            /// Returns an immutable reference to the inner struct.
-            pub fn ffi(&self) -> &[<mjs $ffi_name>] {
-                unsafe { self.0.as_ref().unwrap() }
-            }
-
-            /// Returns a mutable reference to the inner struct.
-            pub unsafe fn ffi_mut(&mut self) -> &mut [<mjs $ffi_name>] {
-                unsafe { self.0.as_mut().unwrap() }
-            }
-
-            /// Returns the message appended to compiler errors.
-            pub fn info(&self) {
-                read_mjs_string(unsafe { self.ffi().info.as_ref().unwrap() });
-            }
-
-            /// Sets the message appended to compiler errors.
-            pub fn set_info(&mut self, info: &str) {
-                write_mjs_string(info, unsafe { self.ffi_mut().info.as_mut().unwrap() });
-            }
-        }
-
-        impl SpecItem for [<Mjs $ffi_name>]<'_> {
-            unsafe fn element_pointer(&self) -> *mut mjsElement {
-                self.ffi().element
-            }
-        }
-
-        // SAFETY: These are safe to implement, as access to them is available only
-        // through methods or through ffi() and ffi_mut() methods, where the latter is unsafe.
-        unsafe impl Sync for [<Mjs $ffi_name>]<'_> {}
-        unsafe impl Send for [<Mjs $ffi_name>]<'_> {}
-    }};
-}
-
-
-/// Implements the userdata method.
-macro_rules! userdata_method {
-    ($type:ty) => {paste::paste!{
-        /// Returns an immutable slice to userdata.
-        pub fn userdata(&self) -> &[$type] {
-            [<read_mjs_vec_ $type>](unsafe { (*self.0).userdata.as_ref().unwrap() })
-        }
-        
-        /// Sets `userdata`.
-        pub fn set_userdata<T: AsRef<[$type]>>(&mut self, userdata: T) {
-            [<write_mjs_vec_ $type>](userdata.as_ref(), unsafe {self.ffi_mut().userdata.as_mut().unwrap() })
-        }
-    }};
-}
-
-/// Implements vector of strings methods for given attribute $name.
-macro_rules! vec_string_set_append {
-    ($($name:ident),*) => {paste::paste!{
-        $(
-            #[doc = concat!("Splits the `", stringify!($name), "` and put the split text as the corresponding attribute.")]
-            pub fn [<set_ $name>](&mut self, $name: &str) {
-                write_mjs_vec_string($name, unsafe { self.ffi_mut().$name.as_mut().unwrap() });
-            }
-
-            #[doc = concat!("Splits the `", stringify!($name), "` and append the split text to the corresponding attribute.")]
-            pub fn [<append_ $name>](&mut self, $name: &str) {
-                append_mjs_vec_string($name, unsafe { self.ffi_mut().$name.as_mut().unwrap() });
-            }
-        )*
-    }};
-}
-
-/// Implements string methods for given attribute $name.
-macro_rules! string_set_get {
-    ($($name:ident),*) => {paste::paste!{
-        $(
-            #[doc = concat!("Returns `", stringify!($name), "`.")]
-            pub fn $name(&self) -> &str {
-                read_mjs_string(unsafe { (*self.0).$name.as_ref().unwrap() })
-            }
-
-            #[doc = concat!("Sets `", stringify!($name), "`.")]
-            pub fn [<set_ $name>](&mut self, $name: &str) {
-                write_mjs_string($name, unsafe { self.ffi_mut().$name.as_mut().unwrap() })
-            }
-        )*
-    }};
-}
 
 
 /***************************
@@ -1017,45 +724,6 @@ impl MjsMaterial<'_> {
 
     vec_string_set_append!(textures);
 }
-
-
-/***************************
-** Default specification
-***************************/
-// This is implemented manually since we can't directly borrow check if something is using the default.
-// We also override the delete method to panic instead of deleting.
-
-/// Default specification. This wraps the FFI type [`mjsDefault`] internally.
-pub struct MjsDefault<'s>(*mut mjsDefault, PhantomData<&'s mut ()>);  // the lifetime belongs to the parent
-
-impl MjsDefault<'_> {
-    /// Returns an immutable reference to the inner struct.
-    pub fn ffi(&self) -> &mjsDefault {
-        unsafe { self.0.as_ref().unwrap() }
-    }
-
-    /// Returns a mutable reference to the inner struct.
-    pub unsafe fn ffi_mut(&mut self) -> &mut mjsDefault {
-        unsafe { self.0.as_mut().unwrap() }
-    }
-}
-
-impl SpecItem for MjsDefault<'_> {
-    unsafe fn element_pointer(&self) -> *mut mjsElement {
-        self.ffi().element
-    }
-
-    /// Defaults can't be deleted.
-    fn delete(self) -> Result<(), Error> {
-        unimplemented!()
-    }
-}
-
-// SAFETY: These are safe to implement, as access to them is available only
-// through methods or through ffi() and ffi_mut() methods, where the latter is unsafe.
-unsafe impl Sync for MjsDefault<'_> {}
-unsafe impl Send for MjsDefault<'_> {}
-
 
 
 /***************************
